producer/app/requirements.txt
black==22.1.0
boto3==1.20.49
botocore==1.23.49
click==8.0.3
jmespath==0.10.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.1
six==1.16.0
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
